requirements.txt
python-dotenv==1.0.1
streamlit==1.36.0
openai>=1.30.0
together>=1.3.0